refactor: removed redundant java package
